add ability to modify document link text label Issue #3
flake8 issue

template name fix

added additional backgrounds and themes

fixed manifest, cleaned up migrations and bumped to 1.0.0 pre release 1

fix for bug when changing out a default logo. The default logo is in static directory, but when a user uploads a logo it is added to the media directory. To allow for the two to work the default them has a function to construct the correct url. A get_readonly_fields was added for default themes, not allowing the user to change out the logos. Users do have the option to create a new custom theme.

flake8 fix

update theme background images